+++ src/solc/standard_json/input/settings/selection/mod.rs
@@ -25,7 +25,7 @@ impl Selection {
     ///
     /// Creates the selection required by our compilation process.
     ///
-    pub fn new_required(pipeline: SolcPipeline) -> Self {
+    pub fn new_required(pipeline: Option<SolcPipeline>) -> Self {
         Self {
             all: Some(FileSelection::new_required(pipeline)),
         }
@@ -34,7 +34,7 @@ impl Selection {
     ///
     /// Extends the user's output selection with flag required by our compilation process.
     ///
-    pub fn extend_with_required(&mut self, pipeline: SolcPipeline) -> &mut Self {
+    pub fn extend_with_required(&mut self, pipeline: Option<SolcPipeline>) -> &mut Self {
         self.all
             .get_or_insert_with(|| FileSelection::new_required(pipeline))
             .extend_with_required(pipeline);