+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- The support for `urls` in standard JSON input
+- The support for `urls` to local files in standard JSON input
 - The support for Yul in standard JSON mode
 
 ## [1.4.1] - 2024-04-24